890cb95b-b126-4e22-a86c-3f102adc4926

+++ mobile/flutter_app/lib/screens/analytics_screen.dart
@@ -186,7 +186,7 @@ class _ProviderBreakdown extends StatelessWidget {
       final existing = providerStats[s.provider] ?? _ProviderStat();
       providerStats[s.provider] = _ProviderStat(
         requests: existing.requests + 1,
-        tokens: existing.tokens + s.totalTokens,
+        tokens: existing.tokens + s.totalTokens as int,
         cost: existing.cost + s.cost,
       );
     }
